Added portfolio mechanic
It is still a bit glitchy and the CSS is all over the place, but it does work. One issue I found is that for some reason, when you press on portfolios, it shows the links for the portfolios of the users, but it shows the description of them below that with all the coding tags, like <li>, <ol>, <p> and any other ones used in the description. I do not know how to solve this, so I would like to know if someone knows how to fix it.